requirements.txt
accelerate==0.21.0
antlr4-python3-runtime==4.9.3
av==14.0.1
Brotli==1.0.9
certifi==2024.12.14
charset-normalizer==3.3.2
diffusers==0.24.0
einops==0.8.0
filelock==3.13.1
fsspec==2024.12.0
gmpy2==2.1.2
huggingface-hub==0.24.0
idna==3.7
imageio==2.36.1
importlib_metadata==8.5.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.3
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.8.1.78
packaging==24.2
pillow==11.0.0
pip==24.2
psutil==6.1.1
PySocks==1.7.1
PyWavelets==1.8.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.21.0
scipy==1.15.1
setuptools==75.1.0
sympy==1.13.3
tifffile==2025.1.10
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.67.1
transformers==4.30.2
triton==2.0.0
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
zipp==3.21.0
